refactor: move business rule mechanism to its own folder
